Reorganize api-collection commands
