+++ src/azure-cli/azure/cli/command_modules/security/aaz/latest/security/api_collection/__init__.py
@@ -9,8 +9,3 @@
 # flake8: noqa
 
 from .__cmd_group import *
-from ._create import *
-from ._delete import *
-from ._list import *
-from ._show import *
-from ._wait import *
