+++ src/azure-cli/azure/cli/command_modules/security/aaz/latest/security/api_collection/apim/_wait.py
@@ -12,7 +12,7 @@ from azure.cli.core.aaz import *
 
 
 @register_command(
-    "security api-collection wait",
+    "security api-collection apim wait",
 )
 class Wait(AAZWaitCommand):
     """Place the CLI in a waiting state until a condition is met.
