+++ src/azure-cli/azure/cli/command_modules/security/tests/latest/test_security_api_collections.py
@@ -17,21 +17,21 @@ class ApiCollectionsTests(ScenarioTest):
 
         nexttoken = base64.b64encode('{"next_link": null, "offset": 0}'.encode()).decode()
 
-        self.cmd("az security api-collection create -g {} --api-id {} --service-name {}".format(resource_group, api_id, service_name), checks=[
+        self.cmd("az security api-collection apim onboard -g {} --api-id {} --service-name {}".format(resource_group, api_id, service_name), checks=[
             JMESPathCheck('name', api_id)
         ])
 
-        self.cmd("az security api-collection wait --created -g {} --api-id {} --service-name {}".format(resource_group, api_id, service_name))
+        self.cmd("az security api-collection apim wait --created -g {} --api-id {} --service-name {}".format(resource_group, api_id, service_name))
 
-        collections = self.cmd("az security api-collection list -g {} --max-items 1 --next-token {}".format(resource_group, nexttoken)).get_output_in_json()
+        collections = self.cmd("az security api-collection apim list -g {} --max-items 1 --next-token {}".format(resource_group, nexttoken)).get_output_in_json()
         assert len(collections) > 0
 
-        collections = self.cmd("az security api-collection list -g {} --service-name {}".format(resource_group, service_name)).get_output_in_json()
+        collections = self.cmd("az security api-collection apim list -g {} --service-name {}".format(resource_group, service_name)).get_output_in_json()
         assert len(collections) > 0
 
-        self.cmd("az security api-collection show -g {} --api-id {} --service-name {}".format(resource_group, api_id, service_name), checks=[
+        self.cmd("az security api-collection apim show -g {} --api-id {} --service-name {}".format(resource_group, api_id, service_name), checks=[
             JMESPathCheck('name', api_id),
             JMESPathCheck('provisioningState', 'Succeeded')
         ])
 
-        self.cmd("az security api-collection delete --yes -g {} --name {} --service-name {}".format(resource_group, api_id, service_name))
+        self.cmd("az security api-collection apim offboard --yes -g {} --name {} --service-name {}".format(resource_group, api_id, service_name))
